XProtect 24R2 release

+++ BoundingBoxMetadataProvider/MainForm.cs
@@ -106,10 +106,16 @@ namespace BoundingBoxMetadataProvider
             if (_metadataProviderChannel != null)
             {
                 _metadataProviderChannel.Disconnect();
+                _metadataProviderChannel.SessionOpening -= MetadataProviderSessionOpening;
+                _metadataProviderChannel.SessionClosed -= MetadataProviderSessionClosed;
                 textBoxSessionCount.Text = @"0";
+            }            
+            if (_metadataProviderService != null)
+            {
+                _metadataProviderService.RemoveMetadataProvider(_metadataProviderChannel);
+                _metadataProviderService.Close();
+                _metadataProviderService = null;
             }
-
-            _metadataProviderService.RemoveMetadataProvider(_metadataProviderChannel);
             buttonConnect.Enabled = true;
             buttonDisconnect.Enabled = false;
             buttonSendData.Enabled = false;
